Custom Theme for Bar Charts

Give some reach colors and fonts to a ggplot.

Theme for dumbbell charts based on theme_default.

Usage

theme_bar(
  flip = TRUE,
  add_text = FALSE,
  axis_text_x_angle = 0,
  axis_text_x_vjust = 0.5,
  axis_text_x_hjust = 0.5
)

theme_default(
  title_font_family = "Carlito",
  title_size = 16,
  title_color = color("dark_grey"),
  title_font_face = "bold",
  title_hjust = NULL,
  title_position_to_plot = TRUE,
  subtitle_font_family = "Carlito",
  subtitle_size = 15,
  subtitle_color = color("dark_grey"),
  subtitle_font_face = "plain",
  subtitle_hjust = NULL,
  text_font_family = "Carlito",
  text_size = 13,
  text_color = color("dark_grey"),
  text_font_face = "plain",
  panel_background_color = "#FFFFFF",
  panel_border = FALSE,
  panel_border_color = color("dark_grey"),
  legend_position = "top",
  legend_direction = "horizontal",
  legend_justification = "center",
  legend_reverse = TRUE,
  legend_title_size = 13,
  legend_title_color = color("dark_grey"),
  legend_title_font_face = "plain",
  legend_title_font_family = "Carlito",
  legend_text_size = 13,
  legend_text_color = color("dark_grey"),
  legend_text_font_face = "plain",
  legend_text_font_family = "Carlito",
  facet_size = 14,
  facet_color = color("dark_grey"),
  facet_font_face = "bold",
  facet_font_family = "Carlito",
  facet_bg_color = color("lighter_grey"),
  axis_x = TRUE,
  axis_y = TRUE,
  axis_text_x = TRUE,
  axis_line_x = FALSE,
  axis_ticks_x = FALSE,
  axis_text_y = TRUE,
  axis_line_y = TRUE,
  axis_ticks_y = TRUE,
  axis_text_font_family = "Carlito",
  axis_text_size = 13,
  axis_text_color = color("dark_grey"),
  axis_text_font_face = "plain",
  axis_title_size = 15,
  axis_title_color = color("dark_grey"),
  axis_title_font_face = "plain",
  axis_text_x_angle = 0,
  axis_text_x_vjust = 0.5,
  axis_text_x_hjust = 0.5,
  grid_major_x = TRUE,
  grid_major_y = FALSE,
  grid_major_color = color("dark_grey"),
  grid_major_x_size = 0.1,
  grid_major_y_size = 0.1,
  grid_minor_x = TRUE,
  grid_minor_y = FALSE,
  grid_minor_color = color("dark_grey"),
  grid_minor_x_size = 0.05,
  grid_minor_y_size = 0.05,
  caption_font_family = "Carlito",
  caption_font_face = "plain",
  caption_position_to_plot = TRUE,
  caption_size = 11,
  caption_color = color("dark_grey"),
  ...
)

theme_dumbbell()

theme_point()

Arguments

flip

Logical. Whether the plot is flipped (horizonal).

axis_text_x_angle

Angle for x-axis text.

axis_text_x_vjust

Vertical justification for x-axis text.

axis_text_x_hjust

Horizontal justification for x-axis text.

title_font_family

Title font family. Default to "Roboto Condensed".

title_size

The size of the legend title. Defaults to 11.

title_color

Legend title color.

title_font_face

Legend title font face. Default to "plain". Font face ("plain", "italic", "bold", "bold.italic").

title_hjust

Title horizontal justification. Default to NULL. Use 0.5 to center the title.

title_position_to_plot

TRUE or FALSE. Positioning to plot or to panel?

text_size

The size of all text other than the title, subtitle and caption. Defaults to 10.

text_color

Text color.

text_font_face

Text font face. Default to "bold". Font face ("plain", "italic", "bold", "bold.italic").

panel_background_color

The color for the panel background color. Default to white.

panel_border

Boolean. Plot a panel border? Default to FALSE.

panel_border_color

A color. Default to REACH main grey.

legend_position

Position of the legend; Default to "right". Can take "right", "left", "top", "bottom" or "none".

legend_direction

Direction of the legend. Default to "vertical". Can take "vertical" or "horizontal".

legend_justification

In addition to legend_direction, place the legend. Can take "left", "bottom", "center", "right", "top".

legend_reverse

Reverse the color in the guide? Default to TRUE.

legend_title_size

Legend title size.

legend_title_color

Legend title color.

legend_title_font_face

Legend title font face. Default to "plain". Font face ("plain", "italic", "bold", "bold.italic").

legend_text_size

Legend text size.

legend_text_color

Legend text color.

legend_text_font_face

Legend text font face. Default to "plain". Font face ("plain", "italic", "bold", "bold.italic").

axis_x

Boolean. Do you need x-axis?

axis_y

Boolean. Do you need y-axis?

axis_text_x

Boolean. Do you need the text for the x-axis?

axis_line_x

Boolean. Do you need the line for the x-axis?

axis_ticks_x

Boolean. Do you need the line for the x-axis?

axis_text_y

Boolean. Do you need the text for the y-axis?

axis_line_y

Boolean. Do you need the line for the y-axis?

axis_ticks_y

Boolean. Do you need the line for the y-axis?

axis_text_size

Axis text size.

axis_text_color

Axis text color.

axis_text_font_face

Axis text font face. Default to "plain". Font face ("plain", "italic", "bold", "bold.italic").

axis_title_size

Axis title size.

axis_title_color

Axis title color.

axis_title_font_face

Axis title font face. Default to "plain". Font face ("plain", "italic", "bold", "bold.italic").

grid_major_x

Boolean. Do you need major grid lines for x-axis?

grid_major_y

Boolean. Do you need major grid lines for y-axis?

grid_major_color

Major grid lines color.

grid_major_x_size

Major X line size.

grid_major_y_size

Major Y line size.

grid_minor_x

Boolean. Do you need minor grid lines for x-axis?

grid_minor_y

Boolean. Do you need minor grid lines for y-axis?

grid_minor_color

Minor grid lines color.

grid_minor_x_size

Minor X line size.

grid_minor_y_size

Minor Y line size.

caption_position_to_plot

TRUE or FALSE. Positioning to plot or to panel?

...

Additional arguments passed to [ggplot2::theme()].

font_family

The font family for all plot's texts. Default to "Segoe UI".

Value

A custom theme object.

A custom theme object.