ggplot2 theme with REACH color palettes

Give some reach colors and fonts to a ggplot.

Usage

theme_reach(
  palette = "main",
  discrete = TRUE,
  reverse = FALSE,
  family = "Leelawadee",
  text_size = 10,
  title_size = 14,
  plot_background_pal = "#FFFFFF",
  panel_background_pal = "#FFFFFF",
  void = FALSE,
  legend_position = "right",
  legend_direction = "vertical",
  legend_reverse = TRUE,
  ...
)

Arguments

palette

Palette name from 'pal_reach()'.

discrete

Boolean indicating whether color aesthetic is discrete or not.

reverse

Boolean indicating whether the palette should be reversed.

family

The font family for all plot's texts. Default to "Leelawadee".

text_size

The size of all text other than the title, subtitle and caption. Defaults to 10.

title_size

The size of the title text_family. Defaults to 14.

plot_background_pal

The color for the plot background color. Default to white.

panel_background_pal

The color for the panel background color. Default to white.

void

Boolean to remove all elements from the plot. Default to FALSE.

legend_position

Position of the legend; Default to "right". Can take "right", "left", "top", "bottom" or "none".

legend_direction

Direction of the legend. Default to "vertical". Can take "vertical" or "horizontal".

legend_reverse

Reverse the color in the guide? Default to TRUE.

...

Additional arguments passed to `ggblanket::gg_theme()`.

Value

The base REACH theme